process stop event

+++ sixsens/process/process.py
@@ -1,3 +1,4 @@
+import time
 import queue
 import multiprocessing
 
@@ -6,6 +7,8 @@ from abc import ABC, abstractmethod
 
 class Process(ABC):
     def __init__(self, *args):
+        self.stop_event = multiprocessing.Event()
+
         self.input_queue, self.output_queue = (
             multiprocessing.Queue(),
             multiprocessing.Queue(),
@@ -13,7 +16,7 @@ class Process(ABC):
 
         self.process = multiprocessing.Process(
             target=self._get_process_function(),
-            args=(self.input_queue, self.output_queue, *args),
+            args=(self.stop_event, self.input_queue, self.output_queue, *args),
         )
         self.process.start()
 
@@ -36,9 +39,12 @@ class Process(ABC):
         return self.latest_data
 
     def stop(self):
-        self.process.terminate()
-        while self.process.is_alive():
-            pass
+        self.stop_event.set()
+
+        time.sleep(0.01)
+
+        if self.process.is_alive():
+            self.process.terminate()
 
         self.process.join(timeout=1.0)
